Improved Tensor hierarchy, update 1.1.

+++ src/waffles/utils/alg/linear/measure/matrix/MatrixOps.java
@@ -64,7 +64,7 @@ public interface MatrixOps extends TensorOps
 		 * 
 		 * @see MatrixOps
 		 */
-		public Qualify(MatrixOps o1, float err)
+		public Qualify(MatrixOps o1, double err)
 		{
 			super(o1, err);
 		}
@@ -143,7 +143,7 @@ public interface MatrixOps extends TensorOps
 
 		
 	@Override
-	public default Operation<Boolean> Allows(float e)
+	public default Operation<Boolean> Allows(double e)
 	{
 		return new Qualify(this, e);
 	}
