+++ src/waffles/utils/alg/linear/measure/matrix/types/shaped/Tall.java
@@ -48,7 +48,7 @@ public interface Tall extends MatrixOps
 		 * 
 		 * @see Tall
 		 */
-		public Qualify(Tall o1, float err)
+		public Qualify(Tall o1, double err)
 		{
 			super(o1, err);
 		}
@@ -82,7 +82,7 @@ public interface Tall extends MatrixOps
 	}
 	
 	@Override
-	public default Operation<Boolean> Allows(float e)
+	public default Operation<Boolean> Allows(double e)
 	{
 		return new Qualify(this, e);
 	}
