+++ src/waffles/utils/alg/linear/measure/matrix/types/shaped/square/Involutory.java
@@ -57,7 +57,7 @@ public interface Involutory extends Square
 		 * 
 		 * @see Involutory
 		 */
-		public Qualify(Involutory o1, float err)
+		public Qualify(Involutory o1, double err)
 		{
 			super(o1, err);
 			
@@ -76,7 +76,7 @@ public interface Involutory extends Square
 		{
 			if(super.result())
 			{
-				float e = Error();
+				double e = Error();
 				Matrix m2 = mp.result();
 				MatrixOps o2 = ID.instance(m2);
 				
@@ -101,7 +101,7 @@ public interface Involutory extends Square
 	}
 	
 	@Override
-	public default Operation<Boolean> Allows(float e)
+	public default Operation<Boolean> Allows(double e)
 	{
 		return new Qualify(this, e);
 	}
